d/hdinsight_cluster: fixing an issue where the `tier` and `kind` were lower-cased due to an issue in the API Spec
The API Specification defines `tier` as a Constant which can be recased - however the `kind` is defined as a string, thus
can't be. I've opened https://github.com/hashicorp/pandora/pull/3404 (to workaround it) and https://github.com/Azure/azure-rest-api-specs/pull/26838
(to fix it) - once those are threaded through both of these should be normalized.